ml-service/requirements.txt
torch==2.2.2+cu121
torchvision==0.17.2+cu121
torchaudio==2.2.2+cu121
--extra-index-url https://download.pytorch.org/whl/cu121
kagglehub
fastapi
uvicorn
python-dotenv
langchain
langchain-core
langchain-community
langchain-groq
langchain-openai
scikit-learn
pandas
matplotlib
numpy==1.26.4
timm
langchain
langchain-core
langchain-community
ddgs
langchain-text-splitters
langchain-openai
langchain-groq
langchain-google-genai 
google-generativeai
langchain-deepseek
streamlit
config